aking changes to my item services

+++ src/main/java/org/olubiyi/mycarauction/dtos/request/AuctionRequestDto.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class AuctionRequestDto {
-    private int reservedPrice;
+    private double reservedPrice;
     private String seller;
     private double startingPrice;
     private String itemId;   // 👈 changed to String (UUID)
